The problem of distinguishing prime numbers from composite numbers and of resolving the latter into their prime factors is known to be one of the most important and useful in arithmetic

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,8 +7,8 @@
  */
 int main(void)
 {
-	long long int i, p_num, div, largest;
-	long long int num = 612852475143;
+	long int i, p_num, div, largest;
+	long int num = 612852475143;
 
 	p_num = num;
 	for (i = 2; i < p_num; i++)
